+++ engine/flink/components/table/src/main/scala/pl/touk/nussknacker/engine/flink/table/aggregate/TableAggregation.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.flink.table.aggregate
 
 import org.apache.flink.api.common.functions.{FlatMapFunction, RuntimeContext}
-import org.apache.flink.api.common.typeinfo.Types
+import org.apache.flink.api.common.typeinfo.{TypeInformation, Types}
 import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.datastream.DataStream
 import org.apache.flink.streaming.api.functions.ProcessFunction
@@ -14,11 +14,13 @@ import pl.touk.nussknacker.engine.api.VariableConstants.KeyVariableName
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.runtimecontext.{ContextIdGenerator, EngineRuntimeContext}
+import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.flink.api.process.{
   AbstractLazyParameterInterpreterFunction,
   FlinkCustomNodeContext,
   FlinkCustomStreamTransformation
 }
+import pl.touk.nussknacker.engine.flink.api.typeinformation.TypeInformationDetection
 import pl.touk.nussknacker.engine.flink.table.aggregate.TableAggregation.{
   aggregateByInternalColumnName,
   groupByInternalColumnName
@@ -34,6 +36,7 @@ class TableAggregation(
     groupByLazyParam: LazyParameter[AnyRef],
     aggregateByLazyParam: LazyParameter[AnyRef],
     selectedAggregator: TableAggregator,
+    aggregationResultType: TypingResult,
     nodeId: NodeId
 ) extends FlinkCustomStreamTransformation
     with Serializable {
@@ -47,7 +50,7 @@ class TableAggregation(
 
     val streamOfRows = start.flatMap(
       new GroupByInputPreparingFunction(groupByLazyParam, aggregateByLazyParam, context),
-      groupByInputTypeInfo(context)
+      GroupByInputPreparingFunction.outputTypeInfo
     )
 
     val inputParametersTable = tableEnv.fromDataStream(streamOfRows)
@@ -64,7 +67,7 @@ class TableAggregation(
     groupedStream
       .process(
         new AggregateResultContextFunction(context.convertToEngineRuntimeContext),
-        aggregateResultTypeInfo(context)
+        AggregateResultContextFunction.outputTypeInfo
       )
   }
 
@@ -93,16 +96,18 @@ class TableAggregation(
 
   }
 
-  private def groupByInputTypeInfo(context: FlinkCustomNodeContext) = {
-    Types.ROW_NAMED(
+  private object GroupByInputPreparingFunction {
+
+    val outputTypeInfo: TypeInformation[Row] = Types.ROW_NAMED(
       Array(groupByInternalColumnName, aggregateByInternalColumnName),
-      context.typeInformationDetection.forType(
+      TypeInformationDetection.instance.forType(
         ToTableTypeEncoder.alignTypingResult(groupByLazyParam.returnType)
       ),
-      context.typeInformationDetection.forType(
+      TypeInformationDetection.instance.forType(
         ToTableTypeEncoder.alignTypingResult(aggregateByLazyParam.returnType)
       )
     )
+
   }
 
   private class AggregateResultContextFunction(convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext)
@@ -129,12 +134,15 @@ class TableAggregation(
 
   }
 
-  private def aggregateResultTypeInfo(context: FlinkCustomNodeContext) = {
-    context.typeInformationDetection.forValueWithContext[AnyRef](
-      ValidationContext.empty
-        .withVariableUnsafe(KeyVariableName, ToTableTypeEncoder.alignTypingResult(groupByLazyParam.returnType)),
-      ToTableTypeEncoder.alignTypingResult(aggregateByLazyParam.returnType)
-    )
+  private object AggregateResultContextFunction {
+
+    val outputTypeInfo: TypeInformation[ValueWithContext[AnyRef]] =
+      TypeInformationDetection.instance.forValueWithContext[AnyRef](
+        validationContext = ValidationContext.empty
+          .withVariableUnsafe(KeyVariableName, ToTableTypeEncoder.alignTypingResult(groupByLazyParam.returnType)),
+        value = ToTableTypeEncoder.alignTypingResult(aggregationResultType)
+      )
+
   }
 
 }