+++ engine/flink/executor/src/main/java/pl/touk/nussknacker/engine/process/util/TimeSerializers.java
@@ -23,7 +23,7 @@ import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.Serializer;
 import com.esotericsoftware.kryo.io.Input;
 import com.esotericsoftware.kryo.io.Output;
-import java.io.Serializable;
+
 import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -38,6 +38,7 @@ import java.time.YearMonth;
 import java.time.ZoneId;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
+
 import org.apache.flink.api.common.ExecutionConfig;
 
 /**
@@ -45,8 +46,8 @@ import org.apache.flink.api.common.ExecutionConfig;
  * <p>
  * Serializers for java.time.*, are added as default serializers if java version is >= 8.
  * <p>
- * Serializers are all private for now because they're not expected to be somehow used/extended/accessed by the user. If there
- * should be a case where this is needed it can be changed - for now the public api should be kept as spall as possible.
+ * Serializers are all public because Flink's class-based serialization init instances of serializer by creating
+ * a new instance
  * <p>
  * Implementation note: All serialization is inspired by oracles java.time.Ser.
  */
@@ -69,19 +70,12 @@ public class TimeSerializers {
         addSerializer(config, Period.class, new PeriodSerializer());
     }
 
-    /**
-     * See description in [[pl.touk.nussknacker.engine.process.util.Serializers]]
-     */
-    private static final long globalVersionUID = 20220222162000L;
-
-    private static <T, Y extends Serializer<T> & Serializable> void addSerializer(ExecutionConfig config, Class<T> klass, Y serializer) {
-      config.getRegisteredTypesWithKryoSerializers().put(klass, new ExecutionConfig.SerializableSerializer<>(serializer));
-      config.getDefaultKryoSerializers().put(klass, new ExecutionConfig.SerializableSerializer<>(serializer));
+    @SuppressWarnings("unchecked")
+    private static <T, Y extends Serializer<T>> void addSerializer(ExecutionConfig config, Class<T> klass, Y serializer) {
+        config.addDefaultKryoSerializer(klass, (Class<? extends Serializer<?>>) serializer.getClass());
     }
 
-    private static class DurationSerializer extends Serializer<Duration> implements Serializable {
-        private static final long serialVersionUID = globalVersionUID;
-
+    public static class DurationSerializer extends Serializer<Duration> {
         {
             setImmutable(true);
         }
@@ -98,9 +92,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class InstantSerializer extends Serializer<Instant> implements Serializable {
-        private static final long serialVersionUID = globalVersionUID;
-
+    public static class InstantSerializer extends Serializer<Instant> {
         {
             setImmutable(true);
         }
@@ -117,8 +109,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class LocalDateSerializer extends Serializer<LocalDate> implements Serializable {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class LocalDateSerializer extends Serializer<LocalDate> {
         {
             setImmutable(true);
         }
@@ -145,8 +136,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class LocalDateTimeSerializer extends Serializer<LocalDateTime> implements Serializable {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class LocalDateTimeSerializer extends Serializer<LocalDateTime> {
         {
             setImmutable(true);
         }
@@ -163,8 +153,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class LocalTimeSerializer extends Serializer<LocalTime> implements Serializable {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class LocalTimeSerializer extends Serializer<LocalTime> {
         {
             setImmutable(true);
         }
@@ -223,8 +212,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class ZoneOffsetSerializer extends Serializer<ZoneOffset> implements Serializable {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class ZoneOffsetSerializer extends Serializer<ZoneOffset> {
         {
             setImmutable(true);
         }
@@ -252,8 +240,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class ZoneIdSerializer extends Serializer<ZoneId> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class ZoneIdSerializer extends Serializer<ZoneId> {
         {
             setImmutable(true);
         }
@@ -276,8 +263,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class OffsetTimeSerializer extends Serializer<OffsetTime> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class OffsetTimeSerializer extends Serializer<OffsetTime> {
         {
             setImmutable(true);
         }
@@ -294,8 +280,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class OffsetDateTimeSerializer extends Serializer<OffsetDateTime> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class OffsetDateTimeSerializer extends Serializer<OffsetDateTime> {
         {
             setImmutable(true);
         }
@@ -314,8 +299,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class ZonedDateTimeSerializer extends Serializer<ZonedDateTime> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class ZonedDateTimeSerializer extends Serializer<ZonedDateTime> {
         {
             setImmutable(true);
         }
@@ -334,8 +318,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class YearSerializer extends Serializer<Year> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class YearSerializer extends Serializer<Year> {
         {
             setImmutable(true);
         }
@@ -349,8 +332,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class YearMonthSerializer extends Serializer<YearMonth> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class YearMonthSerializer extends Serializer<YearMonth> {
         {
             setImmutable(true);
         }
@@ -367,8 +349,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class MonthDaySerializer extends Serializer<MonthDay> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class MonthDaySerializer extends Serializer<MonthDay> {
         {
             setImmutable(true);
         }
@@ -385,8 +366,7 @@ public class TimeSerializers {
         }
     }
 
-    private static class PeriodSerializer extends Serializer<Period> implements Serializable  {
-        private static final long serialVersionUID = globalVersionUID;
+    public static class PeriodSerializer extends Serializer<Period> {
         {
             setImmutable(true);
         }